Made PythonSed to completely work with unicode internally (for Python 2 and 3)
Now PythonSed can deal with ANY multi-byte-encoding or language, as long as the encoding
of all the text files and scripts are the same as provided through the encoding parameter
(use encoding 'latin1' to have PythonSed work with the bare bytes, though ;-)).

Everything coming in is decoded to unicode before processing and encoded again on output.

Passed-in output-streams must accept unicode strings though. Whereas passed-in input
streams my deliver bytes or unicode data. Bytes will be decoded into unicode
automatically using the passed-in encoding.